Conflicts:
	k-core/pom.xml

+++ k-core/src/main/java/org/kframework/utils/BinaryLoader.java
@@ -4,6 +4,7 @@ package org.kframework.utils;
 import java.io.*;
 
 import org.kframework.utils.errorsystem.KExceptionManager;
+
 import com.google.inject.Inject;
 import com.google.inject.Injector;
 import com.google.inject.Singleton;
@@ -18,13 +19,6 @@ public class BinaryLoader {
     // be replaced by serialized objects containing data only.
     private final Injector injector;
 
-    @Inject private static BinaryLoader loader;
-
-    @Deprecated
-    public static BinaryLoader instance() {
-        return loader;
-    }
-
     @Inject
     public BinaryLoader(
             KExceptionManager kem,
@@ -33,17 +27,25 @@ public class BinaryLoader {
         this.injector = injector;
     }
 
-    public void save(String fileName, Object o) throws IOException {
+    public void save(File fileName, Object o) throws IOException {
+        File dir = fileName.getAbsoluteFile().getParentFile();
+        if (!dir.exists() && !dir.mkdirs()) {
+            throw KExceptionManager.criticalError("Could not create directory " + dir);
+        }
         try (OutputStream out = new FileOutputStream(fileName)) {
             save(out, o);
         }
     }
 
-    public void saveOrDie(String fileName, Object o) {
+    public void saveOrDie(File fileName, Object o) {
+        File dir = fileName.getAbsoluteFile().getParentFile();
+        if (!dir.exists() && !dir.mkdirs()) {
+            throw KExceptionManager.criticalError("Could not create directory " + dir);
+        }
         try (OutputStream out = new FileOutputStream(fileName)) {
-            saveOrDie(out, o, fileName);
+            saveOrDie(out, o, fileName.getAbsolutePath());
         } catch (IOException e) {
-            kem.registerCriticalError("Could not write to " + fileName, e);
+            throw KExceptionManager.criticalError("Could not write to " + fileName, e);
         }
     }
 
@@ -51,7 +53,7 @@ public class BinaryLoader {
         try {
             save(out, o);
         } catch (IOException e) {
-            kem.registerCriticalError("Could not write to " + fileName, e);
+            throw KExceptionManager.criticalError("Could not write to " + fileName, e);
         }
     }
 
@@ -66,7 +68,7 @@ public class BinaryLoader {
         }
     }
 
-    public <T> T load(Class<T> cls, String fileName) throws IOException, ClassNotFoundException {
+    public <T> T load(Class<T> cls, File fileName) throws IOException, ClassNotFoundException {
         return cls.cast(load(fileName));
     }
 
@@ -74,13 +76,12 @@ public class BinaryLoader {
         return cls.cast(load(in));
     }
 
-    public <T> T loadOrDie(Class<T> cls, String fileName) {
+    public <T> T loadOrDie(Class<T> cls, File fileName) {
         try (InputStream in = new BufferedInputStream(new FileInputStream(fileName))) {
-            return loadOrDie(cls, in, fileName);
+            return loadOrDie(cls, in, fileName.getAbsolutePath());
         } catch (IOException e) {
-            kem.registerCriticalError("Could not read from " + fileName, e);
+            throw KExceptionManager.criticalError("Could not read from " + fileName, e);
         }
-        return null;
     }
 
     public Object load(InputStream in) throws IOException, ClassNotFoundException {
@@ -92,7 +93,7 @@ public class BinaryLoader {
         }
     }
 
-    public Object load(String fileName) throws IOException, ClassNotFoundException {
+    public Object load(File fileName) throws IOException, ClassNotFoundException {
         try (InputStream in = new BufferedInputStream(new FileInputStream(fileName))) {
             return load(in);
         }
@@ -102,18 +103,17 @@ public class BinaryLoader {
         return loadOrDie(cls, in, "input stream");
     }
 
-    public <T> T loadOrDie(Class<T> cls, InputStream in, String fileName) {
+    private <T> T loadOrDie(Class<T> cls, InputStream in, String fileName) {
 
         try {
             return load(cls, in);
         } catch (ClassNotFoundException e) {
             throw new AssertionError("Something wrong with deserialization", e);
         } catch (ObjectStreamException e) {
-            kem.registerCriticalError("Kompiled definition is out of date with "
+            throw KExceptionManager.criticalError("Kompiled definition is out of date with "
                     + "the latest version of the K tool. Please re-run kompile and try again.", e);
         } catch (IOException e) {
-            kem.registerCriticalError("Could not read from " + fileName, e);
+            throw KExceptionManager.criticalError("Could not read from " + fileName, e);
         }
-        return null;
     }
 }
